Add a Closer to synchronize parent and children go-routine
Add a new Closer struct that implements a CloserRef interface, that
interface is a subset of a context.Context and allow for easier control
of go-routines that are start in a parent and children relationship.

Example:

You have a TCP server that spawn a new go-routines per clients and you
want to close all the go-routines when you are shutting down the server.
You create a Closer and you pass a sub closer, both closer will maintain
a relationship, if you call close on the parent, the `Done()` function  of the
children will unblock and allow you to shutdown gracefully.

If you close a Child with the `Close()` the child will remove itself
from the parent list of children making sure that Close() is only called
once.